refactor(datatable): rename src/data/ to src/datatable/ (refs #298)
