+++ src/components/DataTablePanel.tsx
@@ -24,13 +24,13 @@ import { useStyles2, useTheme2 } from '@grafana/ui';
 import { useApplyTransformation } from 'hooks/useApplyTransformation';
 import React, { useEffect, useMemo, useRef, useState } from 'react';
 import { DatatableOptions } from 'types';
-import { BuildColumnDefs, ConvertDataFrameToDataTableFormat } from 'data/dataHelpers';
-import { ApplyColumnWidthHints } from 'data/columnWidthHints';
-import { buildSearchBarLayout } from 'data/buildSearchBarLayout';
+import { BuildColumnDefs, ConvertDataFrameToDataTableFormat } from 'datatable/dataHelpers';
+import { ApplyColumnWidthHints } from 'datatable/columnWidthHints';
+import { buildSearchBarLayout } from 'datatable/buildSearchBarLayout';
 import { datatableThemedStyles } from './styles';
 import { GetDataTransformerID } from 'pipeline/transformations';
-import { DTColumnType } from 'data/types';
-import { ApplyColumnAliases } from 'data/columnAliasing';
+import { DTColumnType } from 'datatable/types';
+import { ApplyColumnAliases } from 'datatable/columnAliasing';
 
 interface Props extends PanelProps<DatatableOptions> { }
